chore: Move ConversationDomainsStoppedFederatingCellDescription into its own file
